Workaround: In density, temperature, and ffprime restricted Psi_n to values in the range 0...2.
Background: In some cases, the initial conditions lead to Psi_n values in the
first iteration of the first equilibrium determination that are way outside
the reaonable range (e.g., -20.). This can cause floating invalid conditions.
To avoid this, the psi_n values are now explicitly restricted to the range
0...2.

THE CAUSE FOR THE STRANGE PSI_N VALUES SHOULD BE IDENTIFIED AND FIXED LATER!


git-svn-id: https://scm.gforge.inria.fr/svn/aster/jorek2/trunk@219 d232b540-7924-0410-b3cd-8064bb5aaf68