Removing imp. density in diff. term in density eq.
In model501, Eq. 5 is an equation for the global density rho
(all species included). The ion diffusion term should therefore
become div(D*grad(rho-rho_imp)) (instead of div(D*grad(rho))).
This is the object of this commit. Note that presently,
rho_imp is called rhon in the code.