Creating and using new variable mgi_tor_norm
This variable is to be used in subroutine mgi_source
in order to normalize the source in such a way that the
total amount of gas deposited per unit time does not
depend on mgi_deltaphi. At the moment, mgi_tor_norm is
initialized in a simple and temporary way which is to be
replaced by an 1D interpolation in a pre-calculated table.