Use PCG RNG, seed per thread and create sampling routines
New parameter in in_particles file: particle_seed(i) containing the seed
to use for the RNG. If set to 0 (default), use /dev/urandom or the
current time to seed (on the root node) and send this to all processes.
They will then use stream omp_tid + 1000000 * mpi_id so that there is no
overlap possible.

Linking with pcg is done with the mod_pcg_32 module, this could be
expanded to use a 128-bit state to generate 64-bit random numbers.
PCG is slightly altered to include a function generating reals. This is
permitted by the Apache 2 license the code is distributed under.

The mod_sampling module gives an implementation of the box-mueller
method and a transformation for sampling space uniformly in R,Z,phi
coordinates. Using a ziggurat or marsaglia-polar method is something to
be investigated in the future.