Update makefile to fix unnecessary compilation issue
.mod files are not overwritten if they do not change, but make
depends on this information. Touch the .mod files explicitly if they
already exist.

Also create empty rules for .f90, Makefile and Makefile.inc
Make will look for these when parsing .dep/*.d files, and will then look
for Makefile.o et cetera. This costs some time and is better removed.
A double-colon rule has also been made in the generation of the .dep/*.d
files, such that the .f90 files are not checked for prerequisites
anymore.