Change distance comparison to meters, purify neighbours function
The original comparison was in m**2, transform this to meters.
Rewrite some of the distance calculations to use the norm2 function.

Add explicit intents to arguments to allow creating a pure function.
For this the pointer attribute was needed on inb1,2. If this does not
work we can at least try to specify intent(in) on node_list and others.