Alter neighbour checking algorithm
Now check each of the sides of an element in turn.
If one of the nodes matches one on the other side, test the other node
on this side by number as well as by distance to allow for X-point
geometries.

Checked for magnetic axis and x-point, produces correct results for
both.