Create field-line tracing executables phi_zero(_write)
phi_zero takes a particle output file and calculates for each particle
the value of phi at the midplane by following the fieldline from the
particles GC position.
phi_zero_write follows the same fieldline but writes it out as
particle_fieldline for visual inspection.

The fieldline tracer implemented is a simple forward Euler integrator.
The performance of this routine is not too great, it might be desirable
to implement a higher-order integrator here such as Adams-Moulton or RK4.

My tests show that the step size (v*dt) should be <= 0.01 for decent
accuracy with the forward Euler tracer.