Update neighbours calculation, remove transform, update find_RZ_nearby
Instead of using the precalculated transform functions when switching
elements we can now calculate this very simply live with the functions
in mod_neighbours.f90.

The algorithm in find_RZ_nearby has been adjusted to move exactly to the
boundary of its element, and then step over the edge and recalculate in
neighbouring element. This removes the risk of calculating basisfunctions
outside of the element.

We add new unit tests for neighbours calculation, as well as one for the
RTree implementation.