Use mpi instead of mpif.h
Fix many misimplementations which went unnoticed with mpif.h.
for the MPI_OP_CREATE see
https://github.com/mpi-forum/mpi-forum-historic/issues/247
The Debian manpages contain an issue (the op should be a subroutine, not
a function, as also shown by the void type in C.)

We should also have a clear path to upgrade to mpi_f08 now.