Add double-precision particle diagnostics
Specify which ones at the top of mod_particle_diagnostics.f90.
Additionally you can now only write some of the diagnostics with the
optional `only` array argument to particle_diagnostics. This should
contain a list of diagnostics numbers to write. Note that all the
diagnostics are still calculated, so there is not much of a speedup, but
it saves immensely in disk space.