Add Catalyst adaptor and Fortran module interface
This is still very much a WIP so the Catalyst adaptor does not do
anything yet.

Also add a USE_CATALYST makefile variable which defines a preprocessor
macro with the same name. In order to build with Catalyst, one needs to
define the following makefile variables (e.g. in Makefile.inc):
* CATALYSTLIB = -L</path/to/catalyst/lib> -lcatalyst
* CATALYSTINCLUDE = </path/to/include/catalyst-2.0/>