Added pumping capability and determination of I_SOL from j_SOL in jorek2_SOLcurrent
A new example named puff_pump shows how to add neutral pumping in the simulation. The pumping works by imperfectly reflecting kinetic neutral super particles in certain elements on the boundary of the grid. In this way, those wall areas act as a pump. The elements can be chosen and the absorption coefficient can be set. Due to the lack of neutral-neutral collisions, the absorption coefficient does not straightforwardly correspond to a certain pumping speed yet.

The SOL current dI_SOL represented by a fluxline is now determined using the j_SOL, the geometrical knowledge on the wall, and the distribution of starting points of the tracers. In this way, the total current can be easily selected by summing over the tracer starting points of interest. The sanity checks that the j_SOL integrated over the full wall should be 0 is also automatically done. Furthermore, the dI_SOL is corrected for the fact that j_SOL is not a fluxfunction so the flux expansion also changes j_par.

All of this is explained in more detail in the Msc. Thesis of D. Maris

Lastly, the eckstein coefficients are checked for correct input, and the equation the code is based on is properly referenced.