bugfix for error in gfortran compilation
in gfortran, associates are private, so setting an object (in our case
particles) shared if there is an associate pointing to a part of that
object gives an error. This is easily resolved by moving the shared
directive as it did not need to be specified anyway given that in
gfortran we use default(shared).