Fix initial tstep bug
tstep was never set to the initial correct value if tstep_n is used in the
namelist. This means tstep is only updated by the stepper and hence had
the default value before the stepper, which is a problem for any program
wanting to use tstep because it is not intuitive. Hence this fix

Minor changes in phys_module to correct double lines due to merge and
making the explanations of kinetic neutral related things with !< rather
than ! so they show up in code editors