test for NNC implementation
the idea is to test the NNC implementation by having a gas only
simulation and comparing time dependent results to theoretical values.
To do this, this setup uses a nearly straight box (R>>a) in which neutrals are
initialised with a certain temperature distribution. The neutrals on the
left half are tagged 1, and on the right half 2. The particles are
progressed and are collided. This leads to diffusion, which can be
compared against theoretical values.
The density per species, temperature and pressures in the 3 directions are
projected to the FE grid. The next step is to build a custom diagnostic
that has the density per species as a function of R so that this can be
fitted to find the diffusion coefficient.