making output more readable
throwing less errors while keeping diagnostic information
cleanup of timing (most time is spent at 11, i.e. loop over i_pair,
which is indeed where we would expect the most time to be spent, there
it is equally spent on the different aspects of 14, 15, 16 and 17
meaning that it is difficult to get much optimisation from here on. 17
is the timing of the timing itself, which takes quite some time, so
removing the timing will give a noticeable speedup actually)