Removed duplicate -gvec option from convert2vtk.sh
An otion to request that GVEC data be saved to the VTK file was added both in this pull request
and in another pull request that was already merged into develop. This lead to duplicate options
(-gvec and -gvec_field) that do the same thing being present in convert2vtk.sh after the merge.
This commit removes one of them.